pyproject.toml
[tool.poetry]
name = "homomorphic-face-encryption"
version = "0.1.0"
description = "Privacy-preserving facial recognition system"
authors = ["Kartik <kartik@example.com>"]
readme = "README.md"
packages = [{include = "homomorphic_face_encryption", from = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
facenet-pytorch = "*"
mtcnn = "*"
opencv-python = "*"
flask = "*"
flask-jwt-extended = "*"
sqlalchemy = "*"
psycopg2-binary = "*"
numpy = "*"
scikit-learn = "*"
pillow = "*"
# openfhe-python = "*"  # Not available on PyPI, install separately if needed

[tool.poetry.scripts]
app = "homomorphic_face_encryption.app:create_app"

[tool.poetry.group.dev.dependencies]
black = "*"
flake8 = "*"
mypy = "*"
pre-commit = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
